smart crab

+++ oop2_project/Projectile.cpp
@@ -8,7 +8,7 @@ Projectile::Projectile(GameScreen& gameScreen, Direction direction)
 
 void Projectile::init(Direction direction)
 {
-	setSize(BoardObject::getDefaultSize().x * 0.5, BoardObject::getDefaultSize().y * 0.5);
+	setSize(static_cast<int>(BoardObject::getDefaultSize().x * 0.5), static_cast<int>(BoardObject::getDefaultSize().y * 0.5));
 	setDrawPriority(DRAW_PRIORITY);
 	//setDirection(direction);
 }
